requirements.txt
openai>=1.12.0
websockets>=12.0
python-dotenv>=1.0.0
pyaudio>=0.2.11
asyncio
aiofiles>=23.0.0